gen-ai-midterm/requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bitsandbytes==0.42.0
blinker==1.9.0
bs4==0.0.2
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.5.14
datasets==2.21.0
Deprecated==1.3.1
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.121.1
filelock==3.20.0
filetype==1.2.0
firecrawl-py==4.6.0
Flask==3.1.2
Flask-Cors==4.0.2
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==0.3.13
langchain-classic==1.0.0
langchain-community==0.3.13
langchain-core==0.3.63
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langgraph==0.2.76
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.2
langgraph-sdk==0.1.74
langsmith==0.2.11
llama-index==0.8.5.post2
llama-index-core==0.11.23
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.2
numexpr==2.14.1
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
overrides==7.7.0
packaging==23.2
pandas==2.3.3
peft==0.17.1
pillow==12.0.0
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pulsar-client==3.8.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.4.2
scipy==1.15.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.49.3
streamlit==1.51.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.3.0
torch==2.2.2
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
transformers==4.46.3
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0